+++ packages/gql.client/src/degen.ts
@@ -8,7 +8,7 @@ const config: CodegenConfig = {
       plugins: [
         "typescript",
         "typescript-operations",
-        "typescript-graphql-request",
+        "typescript-react-apollo",
       ],
     },
   },