+++ packages/server/src/schema/exercise/mutations.ts
@@ -2,11 +2,12 @@ import * as grpc from "@grpc/grpc-js";
 import client from "@gymlabs/admin.grpc.client";
 import {
   BooleanType__Output,
-  WorkoutPlanItem__Output,
-  Workout__Output,
+  ExerciseStep__Output,
+  Exercise__Output,
 } from "@gymlabs/admin.grpc.definition";
 import { ZodError } from "zod";
 
+import { Exercise, ExerciseStep } from "./types";
 import { mapNullToUndefined } from "../../lib/mapNullToUndefined";
 import { meta } from "../../lib/metadata";
 import { builder } from "../builder";
@@ -17,11 +18,10 @@ import {
   UnauthenticatedError,
   UnauthorizedError,
 } from "../errors";
-import { Workout, WorkoutPlanItem } from "../workouts/types";
 
 builder.mutationFields((t) => ({
-  createWorkout: t.fieldWithInput({
-    type: Workout,
+  createExercise: t.fieldWithInput({
+    type: Exercise,
     input: {
       organizationId: t.input.string(),
       name: t.input.string(),
@@ -36,12 +36,12 @@ builder.mutationFields((t) => ({
         UnauthorizedError,
       ],
     },
-    resolve: async (query, { input }, args, context) => {
-      if (!args.viewer.isAuthenticated()) throw new UnauthenticatedError();
+    resolve: async (query, { input }, ctx) => {
+      if (!ctx.viewer.isAuthenticated()) throw new UnauthenticatedError();
       try {
-        const workout: Workout__Output = await new Promise(
+        const exercise: Exercise__Output = await new Promise(
           (resolve, reject) => {
-            client.createWorkout(input, meta(args.viewer), (err, res) => {
+            client.createExercise(input, meta(ctx.viewer), (err, res) => {
               if (err) {
                 reject(err);
               } else if (res) {
@@ -51,14 +51,14 @@ builder.mutationFields((t) => ({
           }
         );
         return {
-          ...workout,
-          items: workout.items.map((item) => ({
-            ...item,
-            createdAt: new Date(item.createdAt),
-            updatedAt: new Date(item.updatedAt),
+          ...exercise,
+          steps: exercise.steps.map((step) => ({
+            ...step,
+            createdAt: new Date(step.createdAt),
+            updatedAt: new Date(step.updatedAt),
           })),
-          createdAt: new Date(workout.createdAt),
-          updatedAt: new Date(workout.updatedAt),
+          createdAt: new Date(exercise.createdAt),
+          updatedAt: new Date(exercise.updatedAt),
         };
       } catch (err) {
         const error = err as grpc.ServiceError;
@@ -73,8 +73,8 @@ builder.mutationFields((t) => ({
       }
     },
   }),
-  updateWorkout: t.fieldWithInput({
-    type: Workout,
+  updateExercise: t.fieldWithInput({
+    type: Exercise,
     input: {
       id: t.input.string(),
       name: t.input.string({ required: false }),
@@ -83,21 +83,21 @@ builder.mutationFields((t) => ({
     errors: {
       types: [
         ZodError,
+        InternalServerError,
         InvalidArgumentError,
         NotFoundError,
-        InternalServerError,
         UnauthenticatedError,
         UnauthorizedError,
       ],
     },
-    resolve: async (query, { input }, args, context) => {
-      if (!args.viewer.isAuthenticated()) throw new UnauthenticatedError();
+    resolve: async (query, { input }, ctx) => {
+      if (!ctx.viewer.isAuthenticated()) throw new UnauthenticatedError();
       try {
-        const workout: Workout__Output = await new Promise(
+        const exercise: Exercise__Output = await new Promise(
           (resolve, reject) => {
-            client.updateWorkout(
+            client.updateExercise(
               mapNullToUndefined(input),
-              meta(args.viewer),
+              meta(ctx.viewer),
               (err, res) => {
                 if (err) {
                   reject(err);
@@ -109,22 +109,22 @@ builder.mutationFields((t) => ({
           }
         );
         return {
-          ...workout,
-          items: workout.items.map((item) => ({
-            ...item,
-            createdAt: new Date(item.createdAt),
-            updatedAt: new Date(item.updatedAt),
+          ...exercise,
+          steps: exercise.steps.map((step) => ({
+            ...step,
+            createdAt: new Date(step.createdAt),
+            updatedAt: new Date(step.updatedAt),
           })),
-          createdAt: new Date(workout.createdAt),
-          updatedAt: new Date(workout.updatedAt),
+          createdAt: new Date(exercise.createdAt),
+          updatedAt: new Date(exercise.updatedAt),
         };
       } catch (err) {
         const error = err as grpc.ServiceError;
         switch (error.code) {
-          case grpc.status.NOT_FOUND:
-            throw new NotFoundError(error.message);
           case grpc.status.INVALID_ARGUMENT:
             throw new InvalidArgumentError(error.message);
+          case grpc.status.NOT_FOUND:
+            throw new NotFoundError(error.message);
           case grpc.status.PERMISSION_DENIED:
             throw new UnauthorizedError();
           default:
@@ -133,7 +133,7 @@ builder.mutationFields((t) => ({
       }
     },
   }),
-  deleteWorkout: t.fieldWithInput({
+  deleteExercise: t.fieldWithInput({
     type: "Boolean",
     input: {
       id: t.input.string(),
@@ -148,12 +148,12 @@ builder.mutationFields((t) => ({
         UnauthorizedError,
       ],
     },
-    resolve: async (query, { input }, args, context) => {
-      if (!args.viewer.isAuthenticated()) throw new UnauthenticatedError();
+    resolve: async (query, { input }, ctx) => {
+      if (!ctx.viewer.isAuthenticated()) throw new UnauthenticatedError();
       try {
-        const workout: BooleanType__Output = await new Promise(
+        const exercise: BooleanType__Output = await new Promise(
           (resolve, reject) => {
-            client.deleteWorkout(input, meta(args.viewer), (err, res) => {
+            client.deleteExercise(input, meta(ctx.viewer), (err, res) => {
               if (err) {
                 reject(err);
               } else if (res) {
@@ -162,14 +162,14 @@ builder.mutationFields((t) => ({
             });
           }
         );
-        return workout.value;
+        return exercise.value;
       } catch (err) {
         const error = err as grpc.ServiceError;
         switch (error.code) {
-          case grpc.status.NOT_FOUND:
-            throw new NotFoundError(error.message);
           case grpc.status.INVALID_ARGUMENT:
             throw new InvalidArgumentError(error.message);
+          case grpc.status.NOT_FOUND:
+            throw new NotFoundError(error.message);
           case grpc.status.PERMISSION_DENIED:
             throw new UnauthorizedError();
           default:
@@ -178,46 +178,41 @@ builder.mutationFields((t) => ({
       }
     },
   }),
-  createWorkoutPlanItem: t.fieldWithInput({
-    type: WorkoutPlanItem,
+  createExerciseStep: t.fieldWithInput({
+    type: ExerciseStep,
     input: {
-      workoutId: t.input.string(),
       exerciseId: t.input.string(),
-      repititions: t.input.intList(),
-      weights: t.input.intList(),
       index: t.input.int(),
+      name: t.input.string(),
+      description: t.input.string(),
     },
     errors: {
       types: [
         ZodError,
-        InvalidArgumentError,
         InternalServerError,
+        InvalidArgumentError,
         UnauthenticatedError,
         UnauthorizedError,
       ],
     },
-    resolve: async (query, { input }, args, context) => {
-      if (!args.viewer.isAuthenticated()) throw new UnauthenticatedError();
+    resolve: async (query, { input }, ctx) => {
+      if (!ctx.viewer.isAuthenticated()) throw new UnauthenticatedError();
       try {
-        const workoutPlanItem: WorkoutPlanItem__Output = await new Promise(
+        const exerciseStep: ExerciseStep__Output = await new Promise(
           (resolve, reject) => {
-            client.createWorkoutPlanItem(
-              input,
-              meta(args.viewer),
-              (err, res) => {
-                if (err) {
-                  reject(err);
-                } else if (res) {
-                  resolve(res);
-                }
+            client.createExerciseStep(input, meta(ctx.viewer), (err, res) => {
+              if (err) {
+                reject(err);
+              } else if (res) {
+                resolve(res);
               }
-            );
+            });
           }
         );
         return {
-          ...workoutPlanItem,
-          createdAt: new Date(workoutPlanItem.createdAt),
-          updatedAt: new Date(workoutPlanItem.updatedAt),
+          ...exerciseStep,
+          createdAt: new Date(exerciseStep.createdAt),
+          updatedAt: new Date(exerciseStep.updatedAt),
         };
       } catch (err) {
         const error = err as grpc.ServiceError;
@@ -232,32 +227,32 @@ builder.mutationFields((t) => ({
       }
     },
   }),
-  updateWorkoutPlanItem: t.fieldWithInput({
-    type: WorkoutPlanItem,
+  updateExerciseStep: t.fieldWithInput({
+    type: ExerciseStep,
     input: {
       id: t.input.string(),
-      repititions: t.input.intList({ required: false }),
-      weights: t.input.intList({ required: false }),
+      name: t.input.string({ required: false }),
+      description: t.input.string({ required: false }),
       index: t.input.int({ required: false }),
     },
     errors: {
       types: [
         ZodError,
+        InternalServerError,
         InvalidArgumentError,
         NotFoundError,
-        InternalServerError,
         UnauthenticatedError,
         UnauthorizedError,
       ],
     },
-    resolve: async (query, { input }, args, context) => {
-      if (!args.viewer.isAuthenticated()) throw new UnauthenticatedError();
+    resolve: async (query, { input }, ctx) => {
+      if (!ctx.viewer.isAuthenticated()) throw new UnauthenticatedError();
       try {
-        const workoutPlanItem: WorkoutPlanItem__Output = await new Promise(
+        const exerciseStep: ExerciseStep__Output = await new Promise(
           (resolve, reject) => {
-            client.updateWorkoutPlanItem(
+            client.updateExerciseStep(
               mapNullToUndefined(input),
-              meta(args.viewer),
+              meta(ctx.viewer),
               (err, res) => {
                 if (err) {
                   reject(err);
@@ -269,17 +264,17 @@ builder.mutationFields((t) => ({
           }
         );
         return {
-          ...workoutPlanItem,
-          createdAt: new Date(workoutPlanItem.createdAt),
-          updatedAt: new Date(workoutPlanItem.updatedAt),
+          ...exerciseStep,
+          createdAt: new Date(exerciseStep.createdAt),
+          updatedAt: new Date(exerciseStep.updatedAt),
         };
       } catch (err) {
         const error = err as grpc.ServiceError;
         switch (error.code) {
-          case grpc.status.NOT_FOUND:
-            throw new NotFoundError(error.message);
           case grpc.status.INVALID_ARGUMENT:
             throw new InvalidArgumentError(error.message);
+          case grpc.status.NOT_FOUND:
+            throw new NotFoundError(error.message);
           case grpc.status.PERMISSION_DENIED:
             throw new UnauthorizedError();
           default:
@@ -288,7 +283,7 @@ builder.mutationFields((t) => ({
       }
     },
   }),
-  deleteWorkoutPlanItem: t.fieldWithInput({
+  deleteExerciseStep: t.fieldWithInput({
     type: "Boolean",
     input: {
       id: t.input.string(),
@@ -297,38 +292,34 @@ builder.mutationFields((t) => ({
       types: [
         ZodError,
         InvalidArgumentError,
-        NotFoundError,
         InternalServerError,
+        NotFoundError,
         UnauthenticatedError,
         UnauthorizedError,
       ],
     },
-    resolve: async (query, { input }, args, context) => {
-      if (!args.viewer.isAuthenticated()) throw new UnauthenticatedError();
+    resolve: async (query, { input }, ctx) => {
+      if (!ctx.viewer.isAuthenticated()) throw new UnauthenticatedError();
       try {
-        const workoutPlanItem: BooleanType__Output = await new Promise(
+        const exerciseStep: BooleanType__Output = await new Promise(
           (resolve, reject) => {
-            client.deleteWorkoutPlanItem(
-              input,
-              meta(args.viewer),
-              (err, res) => {
-                if (err) {
-                  reject(err);
-                } else if (res) {
-                  resolve(res);
-                }
+            client.deleteExerciseStep(input, meta(ctx.viewer), (err, res) => {
+              if (err) {
+                reject(err);
+              } else if (res) {
+                resolve(res);
               }
-            );
+            });
           }
         );
-        return workoutPlanItem.value;
+        return exerciseStep.value;
       } catch (err) {
         const error = err as grpc.ServiceError;
         switch (error.code) {
-          case grpc.status.NOT_FOUND:
-            throw new NotFoundError(error.message);
           case grpc.status.INVALID_ARGUMENT:
             throw new InvalidArgumentError(error.message);
+          case grpc.status.NOT_FOUND:
+            throw new NotFoundError(error.message);
           case grpc.status.PERMISSION_DENIED:
             throw new UnauthorizedError();
           default: